# Refactor Trie hasher using onRoot flag