--- a/storage/statedb/trie.go
+++ b/storage/statedb/trie.go
@@ -539,7 +539,7 @@ func (t *Trie) hashRoot(db *Database, onleaf LeafCallback) (node, node) {
 	}
 	h := newHasher(onleaf)
 	defer returnHasherToPool(h)
-	return h.hashRoot(t.root, db, true)
+	return h.hash(t.root, db, true, true)
 }
 
 func GetHashAndHexKey(key []byte) ([]byte, []byte) {